+++ src/pages/Homepage.jsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Header from '../components/header/Header';
+import HeroSection from '../components/hero/HeroSection';
+import Footer from '../components/footer/Footer';
+
+const Homepage = () => {
+    return (
+        <div>
+            <Header />
+            <HeroSection />
+            <Footer />
+        </div>
+    );
+};
+
+export default Homepage;
